% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{meanOfVariable}
\alias{meanOfVariable}
\title{Function meanOfVariable}
\source{
Andy Field, 'Discovering Statistics Using R', p. 228
}
\usage{
meanOfVariable(variable)
}
\arguments{
\item{variable}{the variable to calculate the mean value of.}
}
\value{
None
}
\description{
This function calculates the mean value of a variable and prints the result
}
\examples{
meanOfVariable(c(1,2,3))
}
