% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{rcontrast}
\alias{rcontrast}
\title{Function rcontrast}
\usage{
rcontrast(t, df)
}
\arguments{
\item{t}{t-value of the test.}

\item{df}{degrees of freedom.}
}
\value{
None
}
\description{
DSUR, page 457.
This function computes effect sizes for the orthogonal contrasts.
}
\examples{
rcontrast(2.474, 12)
}
