% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{residual.stats}
\alias{residual.stats}
\title{Function residual.stats}
\source{
Andy Field, 'Discovering Statistics Using R', p. 785
}
\usage{
residual.stats(matrix)
}
\arguments{
\item{matrix}{residual matrix. (Calculate matrix using factor.residuals()).}
}
\value{
None
}
\description{
This function wraps all of the factor analysis residual commands and prints the results.
}
