% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{rFromWilcox}
\alias{rFromWilcox}
\title{Function rFromWilcox}
\source{
Field, Andy. (2012) 'Discovering Statistics Using R', p. 665. London, England: SAGE.
}
\usage{
rFromWilcox(wilcoxModel, N)
}
\arguments{
\item{wilcoxModel}{a model created using wilcox.test().}

\item{N}{total sample size.}
}
\description{
This function computes effect sizes from models created using wilcox.test() and prints the result.
}
\note{
Permission to publish under MIT license granted by Andy Field. Copyright remains with the authors of the function.
}
\author{
Field, Andy. Miles, Jeremy. Field, Zoë.
}

