% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{rmMeanAdjust}
\alias{rmMeanAdjust}
\title{Function rmMeanAdjust}
\usage{
rmMeanAdjust(dataframe)
}
\arguments{
\item{dataframe}{2-dimensional dataframe with the scores that need adjustment.}
}
\value{
output table of variables and adjustment.
}
\description{
DSUR, page 365.
This function adjusts scores for the fact they were from a repeated-measures design.
}
\examples{
rmMeanAdjust(data.frame(c(1,2,3), c(1,2,3)))
}
