% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{ttestfromMeans}
\alias{ttestfromMeans}
\title{Function ttestfromMeans}
\source{
Andy Field, 'Discovering Statistics Using R', p. 375
}
\usage{
ttestfromMeans(x1, x2, sd1, sd2, n1, n2)
}
\arguments{
\item{x1}{mean of group 1.}

\item{x2}{mean of group 2.}

\item{sd1}{standard deviation of group 1.}

\item{sd2}{standard deviation of group 2.}

\item{n1}{sample size of group 1.}

\item{n2}{sample size of group 2.}
}
\value{
None
}
\description{
This function calculates a t-test for the means, standard deviations, and sample sizes of two groups, and prints the results.
}
