% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{kmo}
\alias{kmo}
\title{KMO Kaiser-Meyer-Olkin Measure of Sampling Adequacy}
\usage{
kmo(data)
}
\arguments{
\item{data}{dataframe to compute the kmo on.}
}
\value{
None
}
\description{
DSUR, page 776.
Function by G. Jay Kerns, Youngstown State University
see http://tolstoy.newcastle.edu.au/R/e2/help/07/08/22816/html
This function computes Kaiser-Meyer-Olkin Measure of Sampling Adequacy.
Needs the MASS package.
}
\examples{
kmo(data.frame(c(1,2,3), c(1,2,3), c(1,2,3)))
}
