% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{kmo}
\alias{kmo}
\title{KMO Kaiser-Meyer-Olkin Measure of Sampling Adequacy}
\source{
Andy Field, 'Discovering Statistics Using R', p. 776. G. Jay Kerns, http://tolstoy.newcastle.edu.au/R/e2/help/07/08/22816.html
}
\usage{
kmo(data)
}
\arguments{
\item{data}{dataframe to compute the kmo on.}
}
\value{
list of the calculated results.
}
\description{
Function based on an implementation by G. Jay Kerns.
This function computes Kaiser-Meyer-Olkin Measure of Sampling Adequacy.
Needs the MASS package.
}
