% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{logisiticPseudoR2s}
\alias{logisiticPseudoR2s}
\title{Function logisticsPseudoR2s}
\source{
Andy Field, 'Discovering Statistics Using R', p. 332
}
\usage{
logisiticPseudoR2s(logModel)
}
\arguments{
\item{logModel}{logistic regression model}
}
\value{
None
}
\description{
This function calculates the various values of R^2 from a logistic regression model and prints the results.
}
