% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{logisiticPseudoR2s}
\alias{logisiticPseudoR2s}
\title{Function logisticsPseudoR2s}
\source{
Field, Andy. (2012) 'Discovering Statistics Using R', p. 332. London, England: SAGE.
}
\usage{
logisiticPseudoR2s(logModel)
}
\arguments{
\item{logModel}{logistic regression model}
}
\value{
None
}
\description{
This function calculates the various values of R^2 from a logistic regression model and prints the results.
}
\note{
Permission to publish under MIT license granted by Andy Field. Copyright remains with the authors of the function.
}
\author{
Field, Andy. Miles, Jeremy. Field, Zoë.
}

