% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{meanOfVariable}
\alias{meanOfVariable}
\title{Function meanOfVariable}
\usage{
meanOfVariable(variable)
}
\arguments{
\item{variable}{the variable to calculate the mean value of.}
}
\description{
DSUR, page 228.
This function calculates the mean value of a variable.
}
\examples{
meanOfVariable(lecturerFriends)
}
