% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{meanOfVariable}
\alias{meanOfVariable}
\title{Function meanOfVariable}
\source{
Field, Andy. (2012) 'Discovering Statistics Using R', p. 228. London, England: SAGE.
}
\usage{
meanOfVariable(variable)
}
\arguments{
\item{variable}{the variable to calculate the mean value of.}
}
\value{
None
}
\description{
This function calculates the mean value of a variable and prints the result
}
\note{
Permission to publish under MIT license granted by Andy Field. Copyright remains with the authors of the function.
}
\examples{
meanOfVariable(c(1,2,3))
}
\author{
Field, Andy. Miles, Jeremy. Field, Zoë.
}

