% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{rcontrast}
\alias{rcontrast}
\title{Function rcontrast}
\source{
Field, Andy. (2012) 'Discovering Statistics Using R', p. 457. London, England: SAGE.
}
\usage{
rcontrast(t, df)
}
\arguments{
\item{t}{t-value of the test.}

\item{df}{degrees of freedom.}
}
\value{
None
}
\description{
This function computes effect sizes for the orthogonal contrasts and prints the result.
}
\examples{
rcontrast(2.474, 12)
}
\author{
Field, Andy. Miles, Jeremy. Field, Zoë.
}

