% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{residual.stats}
\alias{residual.stats}
\title{Function residual.stats}
\usage{
residual.stats(matrix)
}
\arguments{
\item{matrix}{residual matrix. (Calculate using factor.residuals()).}
}
\value{
None
}
\description{
DSUR, page 785.
This function wraps all of the factor analysis residual commands.
}
