% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{rFromWilcox}
\alias{rFromWilcox}
\title{Function rFromWilcox}
\usage{
rFromWilcox(wilcoxModel, N)
}
\arguments{
\item{wilcoxModel}{a model created using wilcox.test().}

\item{N}{total sample size.}
}
\value{
ans list of calculation results
}
\description{
DSUR, page 665.
This function computes effect sizes from models created using wilcox.test().
}
