% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSUR_functions.r
\name{rmMeanAdjust}
\alias{rmMeanAdjust}
\title{Function rmMeanAdjust}
\source{
Field, Andy. (2012) 'Discovering Statistics Using R', p. 365. London, England: SAGE.
}
\usage{
rmMeanAdjust(dataframe)
}
\arguments{
\item{dataframe}{2-dimensional dataframe with the scores that need adjustment.}
}
\value{
output table of variables and adjustment.
}
\description{
This function adjusts scores for the fact they were from a repeated-measures design.
}
\author{
Field, Andy. Miles, Jeremy. Field, Zoë.
}

